requirements.txt
streamlit
requests
bs4
duckdb
parquet
chromadb==0.5.3
datasets==3.1.0
ecologits[litellm,mistralai]==0.5.2
litellm==1.53.5
matplotlib==3.9.2
mistralai==1.2.3
numpy==1.26.4
nltk==3.9.1
pandas==2.2.3
plotly==5.24.1
pydantic==2.10.3
PyMuPDF==1.24.14
python-dotenv==1.0.1
scikit-learn==1.5.2
sentence-transformers==3.3.1
streamlit==1.40.2
tiktoken==0.8.0
tqdm==4.67.1